ページ固有のcomponentはfeatures配下に持ってくる
