+++ frontend/src/features/RegisterShop/index.d.ts
